Asignacion de validaciones en base de datos

+++ src/main/java/com/example/demo/private_endpoint/modules/animal_module/controllers/AnimalInOutController.java
@@ -2,6 +2,7 @@ package com.example.demo.private_endpoint.modules.animal_module.controllers;
 
 import java.util.List;
 
+import com.example.demo.private_endpoint.modules.cage_module.repositories.CageRepository;
 import org.springframework.web.bind.annotation.*;
 
 import com.example.demo.private_endpoint.DTOs.AnimalMovementData;
@@ -18,6 +19,7 @@ import lombok.RequiredArgsConstructor;
 public class AnimalInOutController {
 
     private final AnimalInOutService animalInOutService;
+    private final CageRepository cageRepository;
 
 //    @PostMapping(path = "/register/die/{id_asig}/user/{id_cage}/cage")
 //    public Message animalDied(@RequestBody AnimalMovementData animal_data, @PathVariable long id_asig,
@@ -28,6 +30,10 @@ public class AnimalInOutController {
     @PostMapping(path = "/register/remove/{id_asig}/user/{id_cage}/cage/{type_of_movement}")
     public Message removeAnimal(@RequestBody AnimalMovementData animal_input, @PathVariable long id_asig,
             @PathVariable long id_cage, @PathVariable long type_of_movement) {
+        if (cageRepository.findById(id_cage).get().getFeedAnimal().getAnimal_amount() <= 0) {
+            return new Message("No puedes eliminar mas animales");
+        }
+
         if (type_of_movement == 1) {
             return this.animalInOutService.removeAnAnimalRegister(animal_input, id_cage, id_asig);
         }
@@ -42,11 +48,5 @@ public class AnimalInOutController {
     @PostMapping(path = "/set/animals/{id_asig}/user/{id_cage}/cage/{amount}/amount")
     public Message addAnimals(@PathVariable long id_asig, @PathVariable long id_cage, @PathVariable int amount){
         return this.animalInOutService.setAnAnimalInCage(id_cage, id_asig, amount);
-    }
-
-    @PostMapping(path = "register/remove/all/{id_asig}/user/{id_cage}")
-    public Message removeAllAnimals(@RequestBody List<AnimalMovementData> animal_input, @PathVariable long id_asig,
-            @PathVariable long id_cage) {
-        return this.animalInOutService.removeAllAnimalRegister(animal_input, id_cage, id_asig);
     }
 }
